Change how we reclone projects to fix bind mount bug.
Currently, when we reclone a project repo, we delete the project
folder and reclone to it. Unfortunately, when we do this, Docker
loses the bind mount on the container, so when the reclone is
complete, the container can't see the new files e.g.:

mount:
 - ./src/payments/:/var/www/html/
reclone will `rm -rf /src/payments` breaking the mount.

This bug shows itself  when the build scripts say they can't find
a composer.json to install after recloning.

The new code in this patch changes how we reset the project
folder, no longer removing the directory but instead just its
contents, including hidden files, therefore preserving the bind
mount.

Bug: T360946

+++ build/civicrm-buildkit.sh
@@ -5,7 +5,8 @@ CIVICRM_BUILDKIT_SRC_DIR="src/civicrm-buildkit"
 echo
 echo "**** Clone CiviCRM Buildkit"
 if $(ask_reclone "${CIVICRM_BUILDKIT_SRC_DIR}" "CiviCRM Buildkit repo"); then
-    rm -rf ${CIVICRM_BUILDKIT_SRC_DIR}
+    rm -rf "${CIVICRM_BUILDKIT_SRC_DIR:?}"/*
+    find "${CIVICRM_BUILDKIT_SRC_DIR:?}" -mindepth 1 -name '.*' -exec rm -rf {} +
     git clone "https://github.com/civicrm/civicrm-buildkit.git" ${CIVICRM_BUILDKIT_SRC_DIR}
     echo
 fi
